+++ spell-check/package.json
@@ -1,6 +1,6 @@
 {
   "name"    : "spellchecker-demo",
-  "version" : "0.1.0",
+  "version" : "1.1.0",
   "main"    : "main.js",
   "dependencies": {
       "spellchecker": "2.2.0"
